style: make poll_read_exact a general trait
