+++ clash_lib/src/proxy/vmess/vmess_impl/stream.rs
@@ -1,4 +1,4 @@
-use std::{fmt::Debug, mem::MaybeUninit, pin::Pin, task::Poll, time::SystemTime};
+use std::{fmt::Debug, pin::Pin, task::Poll, time::SystemTime};
 
 use aes_gcm::Aes128Gcm;
 use bytes::{BufMut, BytesMut};
@@ -6,7 +6,7 @@ use chacha20poly1305::ChaCha20Poly1305;
 use futures::ready;
 
 use md5::Md5;
-use tokio::io::{AsyncRead, AsyncWrite, AsyncWriteExt, ReadBuf};
+use tokio::io::{AsyncRead, AsyncWrite, AsyncWriteExt};
 
 use crate::{
     common::{
@@ -78,52 +78,13 @@ enum WriteState {
     FlushingData(usize, (usize, usize)),
 }
 
-pub trait ReadExt {
-    fn poll_read_exact(
-        &mut self,
-        cx: &mut std::task::Context,
-        size: usize,
-    ) -> Poll<std::io::Result<()>>;
-    #[allow(unused)]
-    fn get_data(&self) -> &[u8];
-}
+use crate::common::io::{ReadExactBase, ReadExt};
 
-impl<S: AsyncRead + Unpin> ReadExt for VmessStream<S> {
-    // Read exactly `size` bytes into `read_buf`, starting from position 0.
-    fn poll_read_exact(
-        &mut self,
-        cx: &mut std::task::Context,
-        size: usize,
-    ) -> Poll<std::io::Result<()>> {
-        self.read_buf.reserve(size);
-        unsafe { self.read_buf.set_len(size) }
-        loop {
-            if self.read_pos < size {
-                let dst = unsafe {
-                    &mut *((&mut self.read_buf[self.read_pos..size]) as *mut _
-                        as *mut [MaybeUninit<u8>])
-                };
-                let mut buf = ReadBuf::uninit(dst);
-                let ptr = buf.filled().as_ptr();
-                ready!(Pin::new(&mut self.stream).poll_read(cx, &mut buf))?;
-                assert_eq!(ptr, buf.filled().as_ptr());
-                if buf.filled().is_empty() {
-                    return Poll::Ready(Err(std::io::Error::new(
-                        std::io::ErrorKind::UnexpectedEof,
-                        "unexpected eof",
-                    )));
-                }
-                self.read_pos += buf.filled().len();
-            } else {
-                assert!(self.read_pos == size);
-                self.read_pos = 0;
-                return Poll::Ready(Ok(()));
-            }
-        }
-    }
+impl<S: AsyncRead + Unpin> ReadExactBase for VmessStream<S> {
+    type I = S;
 
-    fn get_data(&self) -> &[u8] {
-        self.read_buf.as_ref()
+    fn decompose(&mut self) -> (&mut Self::I, &mut BytesMut, &mut usize) {
+        (&mut self.stream, &mut self.read_buf, &mut self.read_pos)
     }
 }
